hstore bulkload 代码bugfix
